Separate "Variety" schema
