+++ innopoints/schemas/__init__.py
@@ -3,5 +3,6 @@ from .account import *
 from .activity import *
 from .application import *
 from .product import *
+from .variety import *
 
 from .schemas import *  # temporary until moved all schemas
